working

+++ actions.py
@@ -7,9 +7,9 @@ from gamedata import GameObject
 @dataclass
 class DragData:
     dragged_obj: GameObject
-    start_pos: tuple[int, int]
+    start_pos: tuple[float, float]
     start_pos_z: int
-    end_pos: tuple[int, int] | None = None
+    end_pos: tuple[float, float] | None = None
     end_pos_z: int | None = None
